+++ api/cmd/go.mod
@@ -6,6 +6,7 @@ require github.com/nedpals/supabase-go v0.5.0
 
 require (
 	github.com/go-viper/mapstructure/v2 v2.2.1 // indirect
+	github.com/golang-jwt/jwt/v5 v5.3.0
 	github.com/google/go-querystring v1.1.0 // indirect
 	github.com/joho/godotenv v1.5.1
 )
