+++ api/cmd/tasks.go
@@ -4,6 +4,8 @@ import (
 	"errors"
 	"net/http"
 	"strconv"
+
+	"github.com/fleetctrl/fleetctrl-hub/api/pkgs/utils"
 )
 
 type updateTaskPayload struct {
@@ -19,11 +21,11 @@ func getTasksByKey(w http.ResponseWriter, r *http.Request) {
 		Limit(1).
 		Eq("key", key).
 		Execute(&computer); err != nil {
-		_ = writeError(w, http.StatusInternalServerError, err)
+		_ = utils.WriteError(w, http.StatusInternalServerError, err)
 		return
 	}
 	if len(computer) == 0 {
-		_ = writeError(w, http.StatusNotFound, errors.New("not found"))
+		_ = utils.WriteError(w, http.StatusNotFound, errors.New("not found"))
 		return
 	}
 
@@ -33,10 +35,10 @@ func getTasksByKey(w http.ResponseWriter, r *http.Request) {
 		Eq("computer_id", strconv.Itoa(int(computer[0].ID))).
 		Eq("status", "PENDING").
 		Execute(&tasks); err != nil {
-		_ = writeError(w, http.StatusInternalServerError, err)
+		_ = utils.WriteError(w, http.StatusInternalServerError, err)
 		return
 	}
-	writeJSON(w, http.StatusOK, map[string]any{
+	utils.WriteJSON(w, http.StatusOK, map[string]any{
 		"tasks": tasks,
 	})
 }
@@ -45,9 +47,9 @@ func updateTaskStatus(w http.ResponseWriter, r *http.Request) {
 	id := r.PathValue("id")
 	var payload updateTaskPayload
 
-	err := parseJSON(r, &payload)
+	err := utils.ParseJSON(r, &payload)
 	if err != nil {
-		_ = writeError(w, http.StatusBadRequest, err)
+		_ = utils.WriteError(w, http.StatusBadRequest, err)
 	}
 
 	update := map[string]interface{}{"status": payload.Status, "error": payload.Error}
@@ -56,8 +58,8 @@ func updateTaskStatus(w http.ResponseWriter, r *http.Request) {
 		Update(update).
 		Eq("uuid", id).
 		Execute(&updated); err != nil {
-		_ = writeError(w, http.StatusInternalServerError, err)
+		_ = utils.WriteError(w, http.StatusInternalServerError, err)
 		return
 	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
+	utils.WriteJSON(w, http.StatusOK, map[string]string{"status": "ok"})
 }
